PANDO-68: Add tests for NodeEnumerableSerializer
